changed /cart/ routing: removed /cart/* as a route option

+++ server.js
@@ -69,21 +69,11 @@ app.all(function (req,res,next) {
 // PAGE ROUTING
 // =======================================================
 
-
-
 app.get('/$', function(req,res) {
   res.sendFile(__dirname +'/index.html');
 });
 app.get('/home/?$', function(req,res) {
   res.sendFile(__dirname +'/index.html');
-});
-app.get('/cart/?', function(req,res,next) {
-  console.log("WELCOME TO THE CART");
-  var logged_in = is_logged_in(req.session);
-  if (!logged_in || req.body == undefined) {
-    res.status(403).send("Please log in to view the contents of your cart.");
-  }
-  next();
 });
 app.get('/cart/?$', function(req,res) {
   res.sendFile(__dirname +'/cart.html');
@@ -268,26 +258,36 @@ app.get('/shop/all$', function (req, res) {
 
 /** Load all the items in a user's cart */
 app.get('/cart/all/$', function (req, res) {
+  var sent = false;
+  
+  var logged_in = is_logged_in(req.session);
+  if (!logged_in) {
+    sent=true;
+    res.status(403).send("Please log in to view the contents of your cart.");
+    return;
+  }
   
   var id = get_id (req.session.email);
   console.log("Sending the cart of "+req.session.email +":id="+id+"="+req.session.user_id);
-    
-  var query = client.query("SELECT "+item_table+".name, "+cart_table+".size, "+cart_table+".quantity"
-    +" FROM "+ cart_table+", "+item_table
-    +" WHERE "+cart_table+".uid = "+id+" and "+item_table+".iid = "+cart_table+".iid;");
-  var results = [];
-  query.on('row',function(row) {
-    results.push(row);
-  });
-  query.on('error', function(err) {
-    if(err) {
-      console.log("ERROR: error running query", err);
-      res.status(500).send("Bad request: you do not have access to this cart.");
-    }
-  });
-  query.on('end',function() {
-      res.json(results);
-  });
+  
+  if (!sent) {
+    var query = client.query("SELECT "+item_table+".name, "+cart_table+".size, "+cart_table+".quantity"
+      +" FROM "+ cart_table+", "+item_table
+      +" WHERE "+cart_table+".uid = "+id+" and "+item_table+".iid = "+cart_table+".iid;");
+    var results = [];
+    query.on('row',function(row) {
+      results.push(row);
+    });
+    query.on('error', function(err) {
+      if(err) {
+        console.log("ERROR: error running query", err);
+        res.status(500).send("Bad request: you do not have access to this cart.");
+      }
+    });
+    query.on('end',function() {
+        res.json(results);
+    });
+  }
 });
 
 
@@ -297,39 +297,51 @@ app.get('/cart/all/$', function (req, res) {
   TODO: make the quantities update correctly
 */
 app.post('/cart/delete/:iid/:size$', function (req, res) {
+  var sent = false;
+  var logged_in = is_logged_in(req.session);
+  if (!logged_in) {
+    sent=true;
+    res.status(403).send("Please log in to view the contents of your cart.");
+  }
 	
   var id = get_id (req.session.email);
   console.log("Deleting "+req.params.iid+" from the cart of "+req.session.email +":id="+id+"="+req.session.user_id);
   
-  var query = client.query("DELETE FROM "+ cart_table +" only"
-    +" WHERE uid = "+id+" and iid = "+req.params.iid
-    +" and size = \'"+req.params.size+"\';");
-  var deleted = false;
-  query.on('rows', function(err) {
-    deleted = true;
-  });
-  query.on('error', function(err) {
-    if(err) {
-      console.log("Encountered an error while querying the database: "+ err);
-      res.status(500).send("Sorry, you cannot delete items from this cart.");
-    }
-  });
-  query.on('end',function(result) {
-      if (deleted) {
-        res.status(200).send("Item successfully removed from cart.");
-      } else {
-        res.status(404).send("Item could not be removed from the cart.");
+  if (!sent) {
+    var query = client.query("DELETE FROM "+ cart_table +" only"
+      +" WHERE uid = "+id+" and iid = "+req.params.iid
+      +" and size = \'"+req.params.size+"\';");
+    var deleted = false;
+    query.on('rows', function(err) {
+      deleted = true;
+    });
+    query.on('error', function(err) {
+      if(err) {
+        console.log("Encountered an error while querying the database: "+ err);
+        res.status(500).send("Sorry, you cannot delete items from this cart.");
       }
-  });
+    });
+    query.on('end',function(result) {
+        if (deleted) {
+          res.status(200).send("Item successfully removed from cart.");
+        } else {
+          res.status(404).send("Item could not be removed from the cart.");
+        }
+    });
+  }
 });
 
 
 /* CHECK OUT ITEMS. */
 app.post('/cart/checkout/', function (req, res) {
-  
-  var id = get_id (req.session.email);
-  console.log(req.session.email +": "+id);
   var wasSent = false;
+  var id = get_id (req.session.email);
+  
+  var logged_in = is_logged_in(req.session);
+  if (!logged_in) {
+    res.status(403).send("Please log in to view the contents of your cart.");
+  }
+  console.log(req.session.email +": "+id);
       //         [                       date                               ]  [uid][ itemid ] [quantity]                 [orderid]
   var query = client.query("INSERT INTO "+order_table+" (placed,uid,iid,quantity,oid)"
       +" SELECT '"+date.getFullYear()+"-"+date.getMonth()+"-"+date.getDay()+"', uid,    iid,    "+cart_table+".quantity ,   oid"
@@ -487,11 +499,15 @@ app.get('/shop/:category$', function (req, res) {
 // =======================================================
 //app.use('/', router);
 
+
 app.listen(port, function () {
 	console.log('Your app is listening on port ' + port);
 });
 
 function get_id(username) {
+  if (username === undefined || username === null) {
+    return -1;
+  }
 	var query = client.query('select id from users where email = \'' +username +'\';');
 	var id = 0;
 	query.on('row', function(row) {
